Rename platform to platform_impl
