+++ src/lib.rs
@@ -96,7 +96,7 @@ pub use icon::*;
 pub mod dpi;
 mod events;
 mod icon;
-mod platform;
+mod platform_impl;
 mod window;
 
 pub mod os;
@@ -121,7 +121,7 @@ pub mod os;
 /// });
 /// ```
 pub struct Window {
-    window: platform::Window,
+    window: platform_impl::Window,
 }
 
 /// Identifier of a window. Unique for each window.
@@ -131,7 +131,7 @@ pub struct Window {
 /// Whenever you receive an event specific to a window, this event contains a `WindowId` which you
 /// can then compare to the ids of your windows.
 #[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct WindowId(platform::WindowId);
+pub struct WindowId(platform_impl::WindowId);
 
 /// Identifier of an input device.
 ///
@@ -139,7 +139,7 @@ pub struct WindowId(platform::WindowId);
 /// identifies its origin. Note that devices may be virtual (representing an on-screen cursor and keyboard focus) or
 /// physical. Virtual devices typically aggregate inputs from multiple physical devices.
 #[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct DeviceId(platform::DeviceId);
+pub struct DeviceId(platform_impl::DeviceId);
 
 /// Provides a way to retrieve events from the system and from the windows that were registered to
 /// the events loop.
@@ -155,7 +155,7 @@ pub struct DeviceId(platform::DeviceId);
 /// `Window` created from this `EventLoop` _can_ be sent to an other thread, and the
 /// `EventLoopProxy` allows you to wakeup an `EventLoop` from an other thread.
 pub struct EventLoop<T> {
-    events_loop: platform::EventLoop<T>,
+    events_loop: platform_impl::EventLoop<T>,
     _marker: ::std::marker::PhantomData<*mut ()> // Not Send nor Sync
 }
 
@@ -198,7 +198,7 @@ impl<T> EventLoop<T> {
     /// fallback on x11. If this variable is set with any other value, winit will panic.
     pub fn new_user_event() -> EventLoop<T> {
         EventLoop {
-            events_loop: platform::EventLoop::new(),
+            events_loop: platform_impl::EventLoop::new(),
             _marker: ::std::marker::PhantomData,
         }
     }
@@ -241,7 +241,7 @@ impl<T> EventLoop<T> {
 /// Used to wake up the `EventLoop` from another thread.
 #[derive(Clone)]
 pub struct EventLoopProxy<T> {
-    events_loop_proxy: platform::EventLoopProxy<T>,
+    events_loop_proxy: platform_impl::EventLoopProxy<T>,
 }
 
 impl<T> EventLoopProxy<T> {
@@ -279,7 +279,7 @@ pub struct WindowBuilder {
     pub window: WindowAttributes,
 
     // Platform-specific configuration. Private.
-    platform_specific: platform::PlatformSpecificWindowBuilderAttributes,
+    platform_specific: platform_impl::PlatformSpecificWindowBuilderAttributes,
 }
 
 /// Error that can happen while creating a window or a headless renderer.
