+++ src/platform_impl/android/ffi.rs
