+++ src/platform_impl/emscripten/ffi.rs
