+++ src/platform_impl/emscripten/mod.rs
