+++ src/platform_impl/linux/x11/util/format.rs
