+++ src/platform_impl/macos/monitor.rs
