+++ src/platform_impl/macos/util.rs
@@ -3,8 +3,8 @@ use cocoa::base::{id, nil};
 use cocoa::foundation::{NSRect, NSUInteger};
 use core_graphics::display::CGDisplay;
 
-use platform::platform::ffi;
-use platform::platform::window::IdRef;
+use platform_impl::platform::ffi;
+use platform_impl::platform::window::IdRef;
 
 pub const EMPTY_RANGE: ffi::NSRange = ffi::NSRange {
     location: ffi::NSNotFound as NSUInteger,
