+++ src/platform_impl/macos/view.rs
@@ -14,10 +14,10 @@ use objc::declare::ClassDecl;
 use objc::runtime::{Class, Object, Protocol, Sel, BOOL, YES};
 
 use {ElementState, Event, KeyboardInput, MouseButton, WindowEvent, WindowId};
-use platform::platform::events_loop::{DEVICE_ID, event_mods, Shared, to_virtual_key_code};
-use platform::platform::util;
-use platform::platform::ffi::*;
-use platform::platform::window::{get_window_id, IdRef};
+use platform_impl::platform::events_loop::{DEVICE_ID, event_mods, Shared, to_virtual_key_code};
+use platform_impl::platform::util;
+use platform_impl::platform::ffi::*;
+use platform_impl::platform::window::{get_window_id, IdRef};
 
 struct ViewState {
     window: id,
