+++ src/platform_impl/macos/window.rs
@@ -38,9 +38,9 @@ use {
 };
 use CreationError::OsError;
 use os::macos::{ActivationPolicy, WindowExt};
-use platform::platform::{ffi, util};
-use platform::platform::events_loop::{EventLoop, Shared};
-use platform::platform::view::{new_view, set_ime_spot};
+use platform_impl::platform::{ffi, util};
+use platform_impl::platform::events_loop::{EventLoop, Shared};
+use platform_impl::platform::view::{new_view, set_ime_spot};
 use window::MonitorId as RootMonitorId;
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
