+++ src/platform_impl/windows/util.rs
